Move pending presenters to middleware.
