+++ cdx_core/app/controllers/device_messages_controller.rb
@@ -47,7 +47,7 @@ class DeviceMessagesController < ApplicationController
     @page            = 1 if @page < 1
     @order_by        = params["order_by"] || "-device_messages.created_at"
     offset           = (@page - 1) * @page_size
-    @device_messages = Presenters::DeviceMessages.index_view(@messages.order(@order_by).limit(@page_size).offset(offset))
+    @device_messages = DeviceMessages::Presenter.index_view(@messages.order(@order_by).limit(@page_size).offset(offset))
   end
 
   def load_message
