+++ cdx_core/app/middleware/devices/presenter.rb
@@ -1,0 +1,25 @@
+module Devices
+  # Presenter method for Devices
+  class Presenter
+    class << self
+      def index_table(devices)
+        devices.map do |device|
+          {
+            id:              device.uuid,
+            name:            device.name,
+            modelName:       device.device_model.full_name,
+            siteName:        format_site_name(device),
+            viewLink:        Rails.application.routes.url_helpers.device_path(device.id)
+          }
+        end
+      end
+
+      protected
+
+      def format_site_name(device)
+        return '' unless device.site
+        device.site.name
+      end
+    end
+  end
+end
