+++ cdx_core/app/views/device_models/index.haml
@@ -23,6 +23,6 @@
               %h3
                 = I18n.t('device_models.index.device_model', count: @total)
             .box-content
-              = react_component "DeviceModelsIndexTable", deviceModels: Presenters::DeviceModels.index_table(@device_models, current_user), orderBy: @order_by
+              = react_component "DeviceModelsIndexTable", deviceModels: DeviceModels::Presenter.index_table(@device_models, current_user), orderBy: @order_by
               .pagination
                 = render 'shared/pagination'
