+++ cdx_core/app/views/devices/index.haml
@@ -20,7 +20,7 @@
               = link_to devices_path(filters_params.merge(format: :csv)), title: I18n.t('devices.index.download_csv') do
                 %span.icon-download.icon-gray
           .box-content
-            = react_component "DevicesIndexTable", devices: Presenters::Devices.index_table(@devices), title: I18n.t('devices.index.device', count: @total),
+            = react_component "DevicesIndexTable", devices: Devices::Presenter.index_table(@devices), title: I18n.t('devices.index.device', count: @total),
                 orderBy: params["order_by"] || "devices.name"
             .pagination
               = render 'shared/pagination'
