+++ cdx_core/spec/middleware/device_messages/presenter_spec.rb
@@ -1,10 +1,10 @@
 require 'spec_helper'
 
-describe Presenters::DeviceMessages do
+describe DeviceMessages::Presenter do
   let(:institution)  { Institution.make }
   let(:site)         { Site.make institution: institution }
   let(:device_model) { DeviceModel.make institution: institution }
-  let(:device)       { Device.make  institution: institution, site: site, device_model: device_model }
+  let(:device)       { Device.make institution: institution, site: site, device_model: device_model }
 
   describe 'patient_view' do
     before :each do
@@ -18,7 +18,7 @@ describe Presenters::DeviceMessages do
     it 'should return elements formated' do
       expect(described_class.index_view(DeviceMessage.all).first).to eq({
         id:                DeviceMessage.first.id,
-        indexStatus:       {:failed=>"Failed (reprocess)", :link=>Rails.application.routes.url_helpers.reprocess_device_message_path(DeviceMessage.first)},
+        indexStatus:       { failed: 'Failed (reprocess)', link: Rails.application.routes.url_helpers.reprocess_device_message_path(DeviceMessage.first) },
         failureReason:     DeviceMessage.first.index_failure_reason,
         modelName:         DeviceMessage.first.device.device_model.name,
         deviceName:        DeviceMessage.first.device.name,
